Moved constants to /api, as it may be used by widgets sometime in the future.
