+++ src/api/Constants.ts
@@ -1,0 +1,11 @@
+export default {
+  ipcMessages: {
+    GET_WIDGETS: "get-widgets",
+    RECEIVE_WIDGETS: "receive-widgets",
+    UNLOAD_WIDGET: "unload-widget",
+    LOAD_WIDGET: "load-widget",
+    CLOSE_MAIN_WINDOW: "close-main-window",
+    MAXIMIZE_MAIN_WINDOW: "maximize-main-window",
+    MINIMIZE_MAIN_WINDOW: "minimize-main-window",
+  }
+};
