--- src/util/constants.ts
@@ -1,9 +1,0 @@
-export default {
-  GET_WIDGETS: "get-widgets",
-  RECEIVE_WIDGETS: "receive-widgets",
-  UNLOAD_WIDGET: "unload-widget",
-  LOAD_WIDGET: "load-widget",
-  CLOSE_MAIN_WINDOW: "close-main-window",
-  MAXIMIZE_MAIN_WINDOW: "maximize-main-window",
-  MINIMIZE_MAIN_WINDOW: "minimize-main-window",
-};
